[FIX] mail: fix composer style for chat windows
Revision on dd8e95cb1a12603940cc32a3d840b6097515df85

The intent of the commit above was to fix the style of the composer,
which was broken when turning the textarea input into a contenteditable
div. While fixing the style for the composer in Discuss or in the Chatter,
it misplaced the initial position of the caret in a chat window.

This commit fixes the issue on chat windows, while also improving the
previous fix.

Known issues:

    In a chat window with more messages than it could be displayed at once,
    when at the bottom of the thread, the line separator of the composer
    and the thread is off by 1 or 2 pixels (vertically).

    > Fixing this issue breaks a lot the alignment of items in the composer
      of Discuss and Chatter, so this small visual glitch is preferred.