[FIX] stock: symmetric reserve and unreserve
When an internal locations has its `scrapped` field set to True,
reserving will increment the quant's `reserved_quantity` while
unreserving will bypass the reservation mechanism (as in a supplier
location). This leads to a desynchronisation between the
partially_available/available move lines and the quant's
`reserved_quantity`, leading later on to "cannot unreserve more than
available" errors.

To prevent this kind of issues, there's a `should_bypass_reservation`
method set on the location that we *should* use everywhere, it was
implemented by [0]. Sadly, this line was not adapted since its
implementaton in [1].

Co-authored by len-odoo <len@odoo.com>

opw 1837900

[0] 9c7ec07542d7da04f2e503f74ef452d931438b8c
[1] cbb283f24f420ece4b50a66ba3a6efc585f76197