[FIX] calendar: fix recurring events rrule definition
Some versions of the datutil library crash when a 'Z' (specifying 'Zulu' time)
is appended at the end of an rrule.
This is because they decided to be stricter when comparing timezone aware/naive
datetimes with rrules.
This bug appears when trying to create a recurring event on the calendar,
defined with an "until" (a certain date) rule.
We explicitly add the timezone information if the 'Z' is in the rrule.

We also remove the rrule ffield from the view, since it is not used and could
only interfere with the logic of the rrule creation.

related to commit 2eb8ff4f14

opw 1841761