[FIX] website_quote: make 'Pay & Confirm' work without website_sale
When creating a db with only website_quote and payment_paypal you end
up with:

AttributeError: 'payment.transaction' object has no attribute '_check_or_create_sale_tx'

when clicking on 'Pay & Confirm' in an online quote.

This is because various controller methods in website_quote use this
method and it is defined by the sale_payment module.

opw-1847977